- use const on more occasions - reduce signed/unsigend conversions - prefer "const" over "CONST" unless in Tcl interface

+++ nsd/config.c
@@ -651,7 +651,7 @@ NsConfigRead(CONST char *file)
 
     Tcl_Close(NULL, chan);
     data = Tcl_GetStringFromObj(buf, &length);
-    conf = strcpy(ns_malloc(length + 1), data);
+    conf = strcpy(ns_malloc((size_t)length + 1U), data);
     Tcl_DecrRefCount(buf);
 
     return conf;
